Egyptian Used Car Price Prediction
The goal of this project is to analyze these features, explore correlations, and develop machine learning models to predict the selling price of used cars. Various regression models will be trained and evaluated to determine which best predicts car prices based on the available data. The project will also explore feature importance and model optimization techniques.